Alpha 4
This version includes the start of the wind walker rotation, with this
it provides a myChi function which returns 1chi=20units so 5*20=100max
chi.  To use this function simply type in myChi < >DESIREDNUMBER<

Next update will include myChi be divided by 20 to provide a 1-5 number
such as the count of balls found in game for easier calling.

##Planned-
Finalize spell lists - Let's get done with the numbers and in with the
words, organization will be 100x easier.

Spell queue?- This one may be a challenge but I think it will be
possible to get out in Alpha 5.

Before Alpha 5 is released you can expect the source to get much
cleaner, commenting maybe?

Rotation backend functions will be moved to RotationBack.lua in the
valid class/spec *This doesnt matter maybe just RotationBack.lua in
rotations?

Bring back BuffCheck remove from rotation.lua.

Dispell logic?

Target Management for Damage classes?

##Fun to know.
The MistWeaver rotation was able to complete Silver in proving grounds,
it's by no means an accomplishment to do but it's still a minor
challenge for some lower geared players such as myself, once spell quene
is implemented we should see less spam of top function spells such as
effuse.