Alpha 5
We have update a few things.

Currently the spellQue is implemented, but not optimized, or used.  The
spell que will allow users to make a macro to que a spell to cast next
that they deam fit.

We've commented everything~ish

You can now understand most of the code easier than before!

Backend rotation information and functions moved to RotationsBack to
keep the Rotation.lua files cleaner than before.